requirements.txt
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
numpy==2.4.4
pandas==3.0.2
python-dateutil==2.9.0.post0
python-telegram-bot==22.7
requests==2.33.1
six==1.17.0
urllib3==2.6.3
